Revamp sms
1. Move to /pkg/sms
2. Use wire.Struct
