+++ pkg/auth/task/spec/spec.go
@@ -2,8 +2,9 @@ package spec
 
 import (
 	"errors"
+
 	"github.com/skygeario/skygear-server/pkg/core/mail"
-	"github.com/skygeario/skygear-server/pkg/core/sms"
+	"github.com/skygeario/skygear-server/pkg/sms"
 )
 
 const (
