+++ pkg/sms/client.go
@@ -1,11 +1,11 @@
 package sms
 
 import (
-	"github.com/skygeario/skygear-server/pkg/core/config"
+	"github.com/skygeario/skygear-server/pkg/auth/config"
 )
 
 type SendOptions struct {
-	MessageConfig config.SMSMessageConfiguration
+	MessageConfig config.SMSMessageConfig
 	To            string
 	Body          string
 }
