+++ pkg/sms/deps.go
@@ -1,0 +1,11 @@
+package sms
+
+import (
+	"github.com/google/wire"
+)
+
+var DependencySet = wire.NewSet(
+	NewNexmoClient,
+	NewTwilioClient,
+	wire.Struct(new(ClientImpl), "*"),
+)
