+++ pkg/sms/nexmo.go
@@ -2,7 +2,7 @@ package sms
 
 import (
 	nexmo "github.com/njern/gonexmo"
-	"github.com/skygeario/skygear-server/pkg/core/config"
+	"github.com/skygeario/skygear-server/pkg/auth/config"
 	"github.com/skygeario/skygear-server/pkg/core/errors"
 )
 
@@ -12,9 +12,9 @@ type NexmoClient struct {
 	NexmoClient *nexmo.Client
 }
 
-func NewNexmoClient(c *config.NexmoConfiguration) *NexmoClient {
+func NewNexmoClient(c *config.NexmoCredentials) *NexmoClient {
 	var nexmoClient *nexmo.Client
-	if c != nil && c.IsValid() {
+	if c != nil {
 		nexmoClient, _ = nexmo.NewClient(c.APIKey, c.APISecret)
 	}
 	return &NexmoClient{
