+++ pkg/sms/twilio.go
@@ -2,7 +2,7 @@ package sms
 
 import (
 	"github.com/sfreiberg/gotwilio"
-	"github.com/skygeario/skygear-server/pkg/core/config"
+	"github.com/skygeario/skygear-server/pkg/auth/config"
 	"github.com/skygeario/skygear-server/pkg/core/errors"
 )
 
@@ -12,9 +12,9 @@ type TwilioClient struct {
 	TwilioClient *gotwilio.Twilio
 }
 
-func NewTwilioClient(c *config.TwilioConfiguration) *TwilioClient {
+func NewTwilioClient(c *config.TwilioCredentials) *TwilioClient {
 	var twilioClient *gotwilio.Twilio
-	if c != nil && c.IsValid() {
+	if c != nil {
 		twilioClient = gotwilio.NewTwilioClient(c.AccountSID, c.AuthToken)
 	}
